functionality for invincibility power up

+++ src/entity.py
@@ -59,10 +59,11 @@ class Entity(pygame.sprite.Sprite):
                     hit_sprite.kill()
                     hit_sprite.apply_powerup()
                 else:
-                    self.health -= 1
-                    if self.health == 0:
-                        self.game.current_state = GameState.GAME_OVER
-                    else:
-                        [obstacle.kill() for obstacle in self.game.obstacles]
-                        [enemy.kill() for enemy in self.game.enemies]
-                        [projectile.kill() for projectile in self.game.projectiles]
+                    if not self.invincible:
+                        self.health -= 1
+                        if self.health == 0:
+                            self.game.current_state = GameState.GAME_OVER
+                        else:
+                            [obstacle.kill() for obstacle in self.game.obstacles]
+                            [enemy.kill() for enemy in self.game.enemies]
+                            [projectile.kill() for projectile in self.game.projectiles]
